move prop constants to props dir
