+++ chebai_graph/preprocessing/properties/augmented_properties.py
@@ -5,8 +5,8 @@ from rdkit import Chem
 
 from chebai_graph.preprocessing import OneHotEncoder, PropertyEncoder
 from chebai_graph.preprocessing.properties import MolecularProperty
+from chebai_graph.preprocessing.properties.constants import *
 from chebai_graph.preprocessing.reader import RuleBasedFGReader
-from chebai_graph.preprocessing.utils.properties_constants import *
 
 
 class AugmentedBondProperty(MolecularProperty, abc):
