+++ chebai_graph/preprocessing/properties/constants.py
